pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bsort"
version = "0.1.0"
description = "Bottle cap detection and classification system"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "Antigravity", email = "antigravity@example.com" }
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "ultralytics>=8.0.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "PyYAML>=6.0",
    "wandb>=0.15.0",
    "typer>=0.9.0",
    "tqdm>=4.65.0",
    "matplotlib>=3.7.0",
    "scikit-learn>=1.3.0",
]

[project.scripts]
bsort = "bsort.main:app"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pylint>=2.17.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "jupyter>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["bsort*"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.pylint.messages_control]
disable = [
    "C0111",  # missing-docstring (we will enforce google style but pylint can be strict)
    "R0903",  # too-few-public-methods
]
